+++ igris/series/block.cpp
@@ -3,12 +3,12 @@
 
 #include <cassert>
 
-void *igris::series_block::get(int num)
+void *igris::series_block::get(size_t num)
 {
     return (void *)((char *)ptr + (elemsize() * (strt + num)));
 }
 
-const void *igris::series_block::get(int num) const
+const void *igris::series_block::get(size_t num) const
 {
     return (const void *)((char *)ptr + (elemsize() * (strt + num)));
 }
@@ -20,18 +20,24 @@ void *igris::series_block::last()
     return (void *)((char *)ptr + (elemsize() * (fini - 1)));
 }
 
-igris::series_block::series_block(igris::series *parent, void *ptr, int size)
+igris::series_block::series_block(igris::series *parent, void *ptr, size_t size)
     : parent(parent), ptr(ptr), size(size), strt(0), fini(0)
 {
     assert(parent);
 }
 
-bool igris::series_block::has_place() const { return size != fini; }
+bool igris::series_block::has_place() const
+{
+    return size != fini;
+}
 
 void *igris::series_block::emplace()
 {
-    int pos = fini++;
+    size_t pos = fini++;
     return (void *)((char *)ptr + (elemsize() * pos));
 }
 
-int igris::series_block::elemsize() const { return parent->elemsize(); }
+size_t igris::series_block::elemsize() const
+{
+    return parent->elemsize();
+}
