+++ igris/series/field_annotation.cpp
@@ -2,8 +2,11 @@
 #include <igris/series/iterator.h>
 
 igris::series_field_annotation::series_field_annotation(
-    const std::string &machname, const std::string &username, int offset,
-    int size, FieldDataType datatype)
+    const std::string &machname,
+    const std::string &username,
+    size_t offset,
+    size_t size,
+    FieldDataType datatype)
     : machname(machname), username(username), offset(offset), size(size),
       datatype(datatype), u()
 {
@@ -27,7 +30,7 @@ igris::series_field_annotation::expand_numeric(const void *record_pointer) const
     case FieldDataType::I32:
         return *(int32_t *)ptr;
     case FieldDataType::I64:
-        return *(int64_t *)ptr;
+        return static_cast<double>(*(int64_t *)ptr);
     case FieldDataType::U8:
         return *(uint8_t *)ptr;
     case FieldDataType::U16:
@@ -35,7 +38,7 @@ igris::series_field_annotation::expand_numeric(const void *record_pointer) const
     case FieldDataType::U32:
         return *(uint32_t *)ptr;
     case FieldDataType::U64:
-        return *(uint64_t *)ptr;
+        return static_cast<double>(*(uint64_t *)ptr);
     case FieldDataType::Bool:
         return *(bool *)ptr;
     default:
@@ -57,7 +60,10 @@ igris::series_field_annotation::expand(const series_iterator &iterator) const
     return expand_numeric(iterator.pointer());
 }
 
-bool igris::series_field_annotation::is_scatter() const { return u.f.scatter; }
+bool igris::series_field_annotation::is_scatter() const
+{
+    return u.f.scatter;
+}
 
 igris::series_field_annotation &igris::series_field_annotation::scatter(bool en)
 {
