+++ igris/series/iterator.cpp
@@ -77,7 +77,7 @@ const igris::series_block *igris::series_iterator::block() const
     return dlist_entry(block_lnk, series_block, lnk);
 }
 
-igris::series_iterator::series_iterator(dlist_head *block_lnk, int num)
+igris::series_iterator::series_iterator(dlist_head *block_lnk, size_t num)
     : block_lnk(block_lnk), num(num)
 {
 }
